+++ lary-external/src/main/java/cn/lary/external/wk/dto/message/MessageSendDTO.java
@@ -16,15 +16,15 @@ public class MessageSendDTO {
     @JsonProperty("stream_no")
     private String streamNo;
     @JsonProperty("from_uid")
-    private int fromUID;
+    private long fromUID;
     @JsonProperty("channel_id")
-    private int channelID;
+    private long channelID;
     @JsonProperty("channel_type")
     private int channelType;
     //过期时间
     private int expire;
     // 订阅者
-    private List<Integer> subscribers;
+    private List<Long> subscribers;
     // message content
     private byte[] payload;
