+++ lary-module/src/main/java/cn/lary/module/app/entity/LaryChannel.java
@@ -25,14 +25,14 @@ public class LaryChannel implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     private Integer channelType;
 
     /**
      * 用户id
      */
-    private Integer uid;
+    private Long uid;
 
     private Boolean isDelete;
