+++ lary-module/src/main/java/cn/lary/module/event/dto/UserRegisterEventDTO.java
@@ -16,7 +16,7 @@ import java.util.Map;
 @Accessors(chain = true)
 public class UserRegisterEventDTO extends AbstractEventData {
 
-    private int uid;
+    private long uid;
 
     private String phone;
