+++ lary-module/src/main/java/cn/lary/module/group/entity/Group.java
@@ -30,7 +30,7 @@ public class Group implements Serializable {
      * 群id
      */
     @TableId(value = "group_id", type = IdType.AUTO)
-    private Integer groupId;
+    private Long groupId;
 
     /**
      * 群名字
@@ -40,7 +40,7 @@ public class Group implements Serializable {
     /**
      * 群主uid
      */
-    private Integer creator;
+    private Long creator;
 
     /**
      * 群状态
